[Feat] 적 카운트 UI 추가

+++ Assets/ScriptableObject_JHW/MonsterData.cs
@@ -7,6 +7,6 @@ public class MonsterData : ScriptableObject
 {
     [Header("Info")]
     [SerializeField] private string monsterName;
-    [SerializeField] private int monsterHP;
+    [SerializeField] private float monsterHP;
     [SerializeField] private GameObject monsterPrefab;
 }
